More javadoc + renaming
