+++ src/main/java/io/strimzi/kproxy/internal/FilterChainFactory.java
@@ -20,7 +20,16 @@ import java.util.List;
 
 import io.strimzi.kproxy.api.filter.KrpcFilter;
 
-public interface FilterFactory {
+/**
+ * Abstracts the creation of a chain of filter instances, hiding the configuration
+ * required for instantiation at the point at which instances are created.
+ * New instances are created during initialization of a downstream channel.
+ */
+public interface FilterChainFactory {
 
+    /**
+     * Create a new chain of filter instances
+     * @return the new chain.
+     */
     List<KrpcFilter> createFilters();
 }
